requirements.txt
# Core dependencies
python-dotenv==1.0.1
anthropic==0.28.0

# Discord
discord.py==2.3.2

# HTTP requests
requests==2.31.0
aiohttp==3.9.5

# Database
SQLAlchemy==2.0.30
alembic==1.13.1

# Scheduling
APScheduler==3.10.4

# Email
secure-smtplib==0.1.1

# Utilities
python-dateutil==2.9.0
pytz==2024.1

# Logging
colorlog==6.8.2

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1
